Add minihud
This is an experiment in representing trunk failure info in a more
compact way. The basic idea is that successes are not very interesting
to us, so we should focus on presenting mostly failure information.

However, we still want to be able to quickly triage a regression. This
is where the grid is most helpful--it makes it very clear when a job
started failing.

To address this use case, you can easily filter the view by job, so you
can see when a failure started pretty easily. Clicking a job name
automatically sets this filter.